Início da análise sintática, invocando o analisador léxico a todo o momento requisitando o próximo token. Definidas as primeiras regras, onde o programa deve iniciar com 'program', seguido de 'begin'. As exceções foram testadas e são disparadas quando for o caso, no caso referentes à ausência das palavras 'program' e 'begin' no início do programa

+++ src/Token.java
@@ -6,9 +6,9 @@
 public class Token {
 
 	private String content;
-	private int    type;
+	private byte    type;
 	
-	public Token(String content, int type) {
+	public Token(String content, byte type) {
 		setContent(content);
 		setType(type);
 	}
@@ -25,7 +25,7 @@ public class Token {
 		return type;
 	}
 
-	public void setType(int type) {
+	public void setType(byte type) {
 		this.type = type;
 	}
